Create base class for M series and documented it. M100,M210,M600 classes are now empty and inherit from base class, will implement additional or remove methods as supported by the drones next.
