+++ Assets/Scripts/Interaction Scripts/ControllerState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7f3c930cb716449459c64c5dcdf81bcc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
